Revert "Gallery"
