+++ css/master.sass
@@ -152,13 +152,6 @@ ul li
   top: 220px
   margin-left: 15px
 
-.galleryContainer
-  display: flex
-  flex-direction: row
-  flex-wrap: wrap
-  height: auto
-  align-items: center
-  margin-top: 100px
 
 /* High Res Laptop View */
