+++ js/components/Skills.js
@@ -10,7 +10,7 @@ class Skills extends React.Component {
       <div className="skillContainer container">
       <SkillHeader soft={this.props.soft} />
 
-        { this.props.list.map(skill => <Skillcard key={skill.key} icon={skill.icon} name={skill.name} info={skill.info} soft={this.props.soft}/>) }
+        { this.props.list.map(skill => <Skillcard icon={skill.icon} name={skill.name} info={skill.info} soft={this.props.soft}/>) }
 
       </div>
     );
